Move the instance-scalar primitives into their own module
